requirements.txt
# --------- pytorch --------- #
torch>=1.13.1
torchvision>=0.14.1
pytorch-lightning==1.9.2
torchmetrics==0.11.1

# --------- hydra --------- #
hydra-core==1.3.1
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
omegaconf>=2.3.0

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
mediapipe       # face detection pre-trained
# sh            # for running bash commands in some tests (linux/macos only)

# --------- data augmentation --------- #
albumentations>=1.3.0

numpy==1.23.5


